backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
sqlmodel>=0.0.22,<1
pydantic>=2.9,<3
pydantic-settings>=2.6,<3
httpx>=0.28,<1
PyNaCl>=1.5,<2
PyJWT[crypto]>=2.10,<3
bcrypt>=4.2,<5
python-dotenv>=1.0,<2
email-validator>=2.1,<3
mcp>=1.6,<2
a2a-sdk>=1.0,<1.1
